VnE/requirements.txt
numpy>=1.25.2
PySide6~=6.6.1
requests~=2.31.0
freetype-py~=2.4.0
PyOpenGL~=3.1.7